discord-bot/requirements.txt
# Discord Bot Dependencies
discord.py==2.3.2
aiohttp==3.9.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
requests==2.31.0
